Create and display favorites (comic # only)
They are all stored in-memory on client. They only contain the
number of the comic so far.

They are displayed in a line under the favorites button. The X
after each one stands in for deleting a favorite, which will
be how comics will be deleted. I am keeping the standalone
"Delete Fav" button in case I think it may make sense to have that
button conditionally on pages of a favorited comic.